[FIX] 잘못된 프로젝트 위치 변경
