+++ src/main/java/com/shp/FmsApplication.java
@@ -1,0 +1,13 @@
+package com.shp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class FmsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(FmsApplication.class, args);
+	}
+
+}
