+++ src/main/java/com/shp/fms/model/request/BranchRequestBody.java
@@ -1,0 +1,10 @@
+package com.shp.fms.model.request;
+
+import lombok.Data;
+
+@Data
+public class BranchRequestBody {
+	private String name;
+	private String address;
+	private String phoneNumber;
+}
