+++ src/main/java/com/shp/fms/repository/EmployeeRepository.java
@@ -1,0 +1,9 @@
+package com.shp.fms.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.shp.fms.model.entity.Employee;
+
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+
+}
