--- src/main/java/com/shpfms/FmsApplication.java
@@ -1,13 +1,0 @@
-package com.shpfms;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class FmsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(FmsApplication.class, args);
-	}
-
-}
