--- src/main/java/com/shpfms/model/BranchInfo.java
@@ -1,13 +1,0 @@
-package com.shpfms.model;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class BranchInfo {
-	private long branchId;
-	private String name;
-	private String address;
-	private String phoneNumber;
-}
