--- src/main/java/com/shpfms/model/entity/Branch.java
@@ -1,31 +1,0 @@
-package com.shpfms.model.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "BRANCH")
-public class Branch {
-	
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long branchId;
-	
-	@Column(name="NAME", nullable=false)
-	private String name;
-	
-	@Column(name="ADDRESS")
-	private String address;
-	
-	@Column(name="PHONE_NUMBER", nullable=false)
-	private String phoneNumber;
-}
