--- src/main/java/com/shpfms/model/request/BranchRequestBody.java
@@ -1,10 +1,0 @@
-package com.shpfms.model.request;
-
-import lombok.Data;
-
-@Data
-public class BranchRequestBody {
-	private String name;
-	private String address;
-	private String phoneNumber;
-}
