--- src/main/java/com/shpfms/repository/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.shpfms.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.shpfms.model.entity.Member;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-}
